[range]: Move Range structure to the root of crate
* Moved the Range structure and its implementation to the root of
  crate as it will be used by the AddressAllocator and IdAllocator
  as well
