+++ src/lib.rs
@@ -8,7 +8,10 @@
 pub mod allocation_engine;
 mod id_allocator;
 
+use crate::allocation_engine::NodeState;
 pub use crate::id_allocator::IdAllocator;
+use std::cmp::max;
+use std::cmp::min;
 use std::result;
 use thiserror::Error;
 
@@ -39,12 +42,123 @@ pub enum Error {
     /// The range that we try to insert into the interval tree is overlapping
     /// with another node from the tree.
     #[error("Addresses are overlapping.{0:?} intersects with existing {1:?}")]
-    Overlap(allocation_engine::Range, allocation_engine::Range),
+    Overlap(Range, Range),
     /// A node state can be changed just from Free to Allocated, other transitions
     /// are not valid.
     #[error("Invalid state transition for node {0:?} from {1:?} to NodeState::Free")]
-    InvalidStateTransition(allocation_engine::Range, allocation_engine::NodeState),
+    InvalidStateTransition(Range, NodeState),
 }
 
 /// Wrapper over std::result::Result
 pub type Result<T> = result::Result<T, Error>;
+
+/// A closed interval range [start, end] used to describe a memory slot that
+/// will be assigned to a device by the VMM. This structure represents the key
+/// of the Node object in the interval tree implementation.
+#[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Hash, Ord, Debug)]
+pub struct Range {
+    /// Lower boundary of the interval.
+    start: u64,
+    /// Upper boundary of the interval.
+    end: u64,
+}
+
+#[allow(clippy::len_without_is_empty)]
+impl Range {
+    /// Create a new Range object.
+    pub fn new(start: u64, end: u64) -> Result<Self> {
+        if start >= end {
+            return Err(Error::InvalidRange(start, end));
+        }
+        Ok(Range { start, end })
+    }
+
+    /// Get length of the range.
+    pub fn len(&self) -> u64 {
+        self.end - self.start + 1
+    }
+
+    /// Check whether two Range objects overlap with each other.
+    pub fn overlaps(&self, other: &Range) -> bool {
+        max(self.start, other.start) <= min(self.end, other.end)
+    }
+
+    /// Check whether the key is fully covered.
+    pub fn contains(&self, other: &Range) -> bool {
+        self.start <= other.start && self.end >= other.end
+    }
+
+    /// Get the lower boundary of the range.
+    pub fn start(&self) -> u64 {
+        self.start
+    }
+
+    /// Get the upper boundary of the range.
+    pub fn end(&self) -> u64 {
+        self.end
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_new_range() {
+        assert_eq!(Range::new(2, 1).unwrap_err(), Error::InvalidRange(2, 1));
+        assert_eq!(Range::new(1, 1).unwrap_err(), Error::InvalidRange(1, 1));
+    }
+
+    #[test]
+    fn test_range_overlaps() {
+        let range_a = Range::new(1u64, 4u64).unwrap();
+        let range_b = Range::new(4u64, 6u64).unwrap();
+        let range_c = Range::new(2u64, 3u64).unwrap();
+        let range_e = Range::new(5u64, 6u64).unwrap();
+
+        assert!(range_a.overlaps(&range_b));
+        assert!(range_b.overlaps(&range_a));
+        assert!(range_a.overlaps(&range_c));
+        assert!(range_c.overlaps(&range_a));
+        assert!(!range_a.overlaps(&range_e));
+        assert!(!range_e.overlaps(&range_a));
+
+        assert_eq!(range_a.len(), 4);
+    }
+
+    #[test]
+    fn test_range_contain() {
+        let range_a = Range::new(2u64, 6u64).unwrap();
+        assert!(range_a.contains(&Range::new(2u64, 3u64).unwrap()));
+        assert!(range_a.contains(&Range::new(3u64, 4u64).unwrap()));
+        assert!(range_a.contains(&Range::new(5u64, 6u64).unwrap()));
+        assert!(!range_a.contains(&Range::new(1u64, 2u64).unwrap()));
+        assert!(!range_a.contains(&Range::new(1u64, 3u64).unwrap()));
+        assert!(!range_a.contains(&Range::new(1u64, 7u64).unwrap()));
+        assert!(!range_a.contains(&Range::new(7u64, 8u64).unwrap()));
+        assert!(!range_a.contains(&Range::new(6u64, 7u64).unwrap()));
+        assert!(!range_a.contains(&Range::new(7u64, 8u64).unwrap()));
+    }
+
+    #[test]
+    fn test_range_ord() {
+        let range_a = Range::new(1, 4).unwrap();
+        let range_b = Range::new(1, 4).unwrap();
+        let range_c = Range::new(1, 3).unwrap();
+        let range_d = Range::new(1, 5).unwrap();
+
+        assert_eq!(range_a, range_b);
+        assert_eq!(range_b, range_a);
+        assert!(range_a > range_c);
+        assert!(range_c < range_a);
+        assert!(range_a < range_d);
+        assert!(range_d > range_a);
+    }
+
+    #[test]
+    fn test_getters() {
+        let range = Range::new(3, 5).unwrap();
+        assert_eq!(range.start(), 3);
+        assert_eq!(range.end(), 5);
+    }
+}
